added missing collection_exsits before deletion

+++ qdrant_client/migrate/migrate.py
@@ -111,8 +111,8 @@ def _recreate_collection(
     src_collection_info = source_client.get_collection(collection_name)
     src_config = src_collection_info.config
     src_payload_schema = src_collection_info.payload_schema
-
-    dest_client.delete_collection(collection_name)
+    if dest_client.collection_exists(collection_name):
+        dest_client.delete_collection(collection_name)
     dest_client.create_collection(
         collection_name,
         vectors_config=src_config.params.vectors,
